requirements.txt
pyserial==3.5
uflash==2.0.0
